suse: extract_patch: fix -C handling when no CVE is found
if no -r is given and -C does not fiend CVE/bsc refs, we end up with a patch
with no reference, not even the default one.
Fix this by filling patchInfos ref with the default ref in case CVE reference
fails to find something and update the patch file with it.

Signed-off-by: Nicolas Morey <nmorey@suse.com>