suse: gen_ordered_patchlist: Handle refresh patches
Refreshed patches are harder than they should be.
We need to figure out which commit introduced then in kernel-source,
then which commit is actually in the linux tree.
And the right moment to apply them back.

gen_ordered_patchlist now handles all that and provides a toDoList
starting with all the refreshed patches to be reverted, and then
and order list of patches containing both the pending ones and
the refreshed patches to reapply

Signed-off-by: Nicolas Morey <nmorey@suse.com>